added configurable properties

+++ src/main/java/tech/kitucode/banking/repository/CardRepository.java
@@ -10,5 +10,5 @@ public interface CardRepository extends JpaRepository<Card, Long> {
     Optional<Card> findOneByPan(String pan);
     Optional<Card> findOneByCardTypeAndAccountId(CardType cardType, Long accountId);
 
-    Long countByAccountId(Long accountId);
+    Integer countByAccountId(Long accountId);
 }
